+++ Services/ContentConverterToA.cs
@@ -8,7 +8,7 @@ namespace DoTuna
             _origin = origin;
         }
 
-        public override string ConvertContent(string content, int threadId)
+        public override string ConvertContent(string content, ulong threadId)
         {
             if (string.IsNullOrEmpty(content)) return string.Empty;
 
@@ -20,7 +20,7 @@ namespace DoTuna
             content = ConvertGeneralLinks(content);
             return content;
         }
-        protected override string MakeAnchorTag(int threadId, string resNo, string text, bool isExternal)
+        protected override string MakeAnchorTag(ulong threadId, string resNo, string text, bool isExternal)
         {
             var targetAttr = isExternal ? " target=\"_blank\"" : "";
             var relativePath = _origin.GetRelativePathTo(_fileNameMap.Get(threadId));
